Home

DEMC.jl

DEMC.demcz_sample

DEMC.demcz_sample_par

DEMC.demcz_sample — Function.

demczsample(logobj, Zmat, N=4, K=10, Ngeneration=5000, Nblocks=1, blockindex=[1:size(Zmat,2)], epsscale=1e-4*ones(size(Zmat,2)), γ=2.38; prevrun=nothing, verbose = true, print_step=100)

Serial run of DEMC chain. At each generation all chains are updated one after the other.

source

DEMC.demcz_sample_par — Function.

demczsamplepar(logobj, Zmat, N=4, K=10, Ngeneration=5000, Nblocks=1, blockindex=[1:size(Zmat,2)], eps_scale=1e-4*ones(size(Zmat,2)), γ=2.38; prevrun=nothing)

Runs each chain on a separate process - Z is updated simultaenously among all chains running in parallel.

source